requirements-mac.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator>=2.0.0
stripe>=5.0.0
python-dotenv==1.0.0
requests==2.31.0 